+++ rpcs3/Emu/CMakeLists.txt
@@ -455,6 +455,8 @@ target_sources(rpcs3_emu PRIVATE
     RSX/Capture/rsx_capture.cpp
     RSX/Capture/rsx_replay.cpp
     RSX/GL/glutils/buffer_object.cpp
+    RSX/GL/glutils/common.cpp
+    RSX/GL/glutils/image.cpp
     RSX/GL/glutils/ring_buffer.cpp
     RSX/GL/GLCommonDecompiler.cpp
     RSX/GL/GLCompute.cpp
