+++ rpcs3/Emu/RSX/GL/GLHelpers.cpp
@@ -14,18 +14,6 @@ namespace gl
 	capabilities g_driver_caps;
 	const fbo screen{};
 
-	static thread_local bool s_tls_primary_context_thread = false;
-
-	void set_primary_context_thread(bool value)
-	{
-		s_tls_primary_context_thread = value;
-	}
-
-	bool is_primary_context_thread()
-	{
-		return s_tls_primary_context_thread;
-	}
-
 	void flush_command_queue(fence& fence_obj)
 	{
 		fence_obj.check_signaled();
